Attendant CRUD complete

+++ users/views.py
@@ -3,10 +3,15 @@ from .models import Attendant, User
 from .forms import UserForm, AttendantForm
 from django.contrib.auth.hashers import make_password
 
-def attendant(request):
+def all_attendant(request):
     attendants = Attendant.objects.all()
     context = {'attendants': attendants}
     return render(request, 'users/all_attendant.html', context)
+
+def details_attendant(request, id):
+    attendant = Attendant.objects.get(id=id)
+    context = {'attendant': attendant}
+    return render(request, 'users/details_attendant.html', context)
 
 def new_attendant(request):
     if request.method == 'POST':
@@ -32,29 +37,31 @@ def new_attendant(request):
         'attendant_form': attendant_form
     })
 
-def update_attendant(request):
+def update_attendant(request, id):
+    attendant = Attendant.objects.get(user=id)
     if request.method == 'POST':
-        user_form = UserForm(request.POST, instance=request.user)
-        attendant_form = AttendantForm(request.POST, instance=request.user.attendant)
+        user_form = UserForm(request.POST, instance=attendant.user)
+        attendant_form = AttendantForm(request.POST, instance=attendant)
         if user_form.is_valid() and attendant_form.is_valid():
+            post_user = user_form.save(commit=False)
+            post_user.password = make_password(user_form.cleaned_data['password'])
             user_form.save()
             attendant_form.save()
             # messages.success(request, _('Your attendant was successfully updated!'))
-            # return redirect('settings:attendant')
-            # return render(request, 'appointments/index.html')
             return redirect('home')
         # else:
         #     messages.error(request, _('Please correct the error below.'))
     else:
-        user_form = UserForm(instance=request.user)
-        attendant_form = AttendantForm(instance=request.user.attendant)
+        user_form = UserForm(instance=attendant.user)
+        attendant_form = AttendantForm(instance=attendant)
     return render(request, 'users/update_attendant.html', {
         'user_form': user_form,
         'attendant_form': attendant_form
     })
 
 def delete_attendant(request, id):
-    attendant = Attendant.objects.get(id=id)
-    User.objects.filter(id=attendant.user.id).delete()
-    Attendant.objects.filter(id=id).delete()
-    return redirect('attendant')
+    if request.method == 'POST':
+        attendant = Attendant.objects.get(id=id)
+        User.objects.filter(id=attendant.user.id).delete()
+        Attendant.objects.filter(id=id).delete()
+    return redirect('all_attendant')
